+++ zeronexus/__init__.py
@@ -3,6 +3,6 @@
 One Platform. Many Capabilities. Zero Dependence on AI.
 """
 
-__version__ = "2.6.0"
+__version__ = "2.6.1"
 __codename__ = "ZeroNexus"
 __shortname__ = "ZN"